Types

+++ cweb/manager.hpp
@@ -73,7 +73,7 @@ public:
     char        *i2cData;
     
 public:
-    static I2CControl initControl( I2CType type, int file, int command, int param );
+    static I2CControl initControl( I2CType type, int file, int command, long param );
     static I2CControl initControl( I2CType type, int file, int command, char *buffer );
     const char *description();
 };
